Separate fault tolerant and pipelined scheduler implementations
